+++ api_tests/tests/reasonable_adjustments/test_error_cases.py
@@ -1,12 +1,14 @@
 import pytest
 import json
 import requests
+from api_tests.scripts.apigee_api import ApigeeDebugApi
+from api_tests.tests.utils import Utils
 from api_tests.config_files import config
 from assertpy import assert_that
 
 
 @pytest.mark.usefixtures("setup")
-class TestInvalidRequestSuite:
+class TestErrorCaseSuite:
     """ A test suite to verify the correct error messages from an invalid request """
 
     @pytest.mark.errors
@@ -340,6 +342,47 @@ class TestInvalidRequestSuite:
                 'x-request-id': 'test'
             }
         )
+        actual_response = json.loads(response.text)
+
+        # Then
+        assert_that(expected_status_code).is_equal_to(response.status_code)
+        assert_that(actual_response['message_id']).is_not_empty()
+        assert_that(expected_response['error']).is_equal_to_ignoring_case(actual_response['error'])
+        assert_that(expected_response['error_description']).is_equal_to_ignoring_case(actual_response['error_description'])
+
+    @pytest.mark.ods
+    @pytest.mark.errors
+    def test_missing_ods(self, use_internal_testing_internal_dev_without_ods_app, get_token):
+        # Given
+        expected_status_code = 500
+        expected_response = {
+            'error': 'missing ODS',
+            'error_description': 'An internal server error occurred. Missing ODS. Contact us for assistance diagnosing this issue: https://digital.nhs.uk/developer/help-and-support quoting Message ID',
+        }
+
+        # When
+        response = Utils.send_request(self)
+        actual_response = json.loads(response.text)
+        
+        # Then
+        assert_that(expected_status_code).is_equal_to(response.status_code)
+        assert_that(actual_response['message_id']).is_not_empty()
+        assert_that(expected_response['error']).is_equal_to_ignoring_case(actual_response['error'])
+        assert_that(expected_response['error_description']).is_equal_to_ignoring_case(actual_response['error_description'])
+
+    @pytest.mark.asid
+    @pytest.mark.errors
+    def test_missing_asid(self, use_internal_testing_internal_dev_without_asid_app, get_token):
+        # Given
+        expected_status_code = 500
+        expected_response = {
+            'error': 'missing ASID',
+            'error_description': 'An internal server error occurred. Missing ASID. Contact us for assistance diagnosing this issue: https://digital.nhs.uk/developer/help-and-support quoting Message ID',
+        }
+
+        # When
+        response = Utils.send_request(self)
+
         actual_response = json.loads(response.text)
 
         # Then
